refactor: Extract coloured logging helpers
