+++ docs/requirements.txt
@@ -25,7 +25,6 @@ charset-normalizer==3.4.3
 choreographer==1.0.9
 ckzg==2.1.2
 click==8.2.1
-coloredlogs==15.0.1
 comm==0.2.3
 contourpy==1.3.3
 cryptography==45.0.7
@@ -76,7 +75,6 @@ h11==0.16.0
 hexbytes==1.3.1
 httpcore==1.0.9
 httpx==0.28.1
-humanfriendly==10.0
 hypersync-temp==0.10.0
 idna==3.10
 imagesize==1.4.1
